New cohort method to calculate number of days at a status (Recidiviz/recidiviz-data#23125)
## Description of the change

This PR creates a new cohort method to calculate the number of days at a
given status following a cohort start event as described in
[this](https://app.zenhub.com/workspaces/analysis-5f8f1c625afb1c0011c7222a/issues/gh/recidiviz/recidiviz-data/22915)
ticket.

At this point the function outputs a dataframe that has a record for
every cohort start and time index with a value that represents the
number of days spent at the input dataframe status. In order to plot an
aggregated value, the user still needs to do that group by and specify
the aggregation function (mean, median, etc).
[This](https://github.com/Recidiviz/recidiviz-research/blob/agaidus/2023-08-09-dev-work/agaidus/2023-08-08-Demo-Cohort-Time-at-Status.ipynb)
notebook demonstrates use of the function.

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#22915

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 58faa7e0a3624d45fd8ac0d197b51aa619264cff