[Justice Counts] Migrate JC Database to our new GCP projects (Recidiviz/recidiviz-data#23214)
## Description of the change

### Summary

We are in the middle of a long-term project to migrate the Justice
Counts applications to their own GCP projects. The design doc is
[here](https://paper.dropbox.com/doc/Design-Doc-Migrating-Justice-Counts-to-new-GCP-project--B~GPL5bZvvSJgwxo2oNy6SAiAg-FfrVJrgCV3oysgTKaDcLi)
and the Zenhub ticket is
[here](https://github.com/Recidiviz/zenhub-tasks/issues/1247).

This PR includes the code changes necessary to migrate the Justice
Counts SQL database over to the new projects. We will keep the database
in the old Recidiviz projects too, because it contains data from the
Phase 0 state scan that still powers some dashboards in Looker. But it
will not be updated moving forward. We will take a snapshot of the
database and copy it to the new instances in the new GCP projects. Once
the migration is done, the Justice Counts apps will do all read/writes
from the database in the new projects.

In particular, this PR does the following:
* ~Adds a new SchemaType.JUSTICE_COUNTS_V2~
* NEW: Adds `secret_prefix` parameters to a fair bit of our DB
connection code. The reason this is needed is because we currently have
an assumption in our code that there is a 1:1 mapping between SchemaType
and instance URL. We use the SchemaType to determine what secrets to
look at to determine the instance URL. The Justice Counts SchemaType is
now used in two instances, and so there are two sets of secrets, one for
each instance. The first set is prefixed with `justice_counts` and the
second set is prefixed with `justice_counts_v2`. [Note: when Publisher
is migrated to the new JC GCP project, we'll be able to get rid of this.
The Recidiviz GCP project will have one set of secrets that point to the
original JC instance, and the JC GCP projects will have one set of
secrets that point to the new JC instance.]
* Updates the run_migrations_to_head.py script to be able to run
migrations against the new instance
* Updates the access_cloudsql script to be able to connect to the new
instance
* Updates the JC deploy scripts to run migrations against the new
instance
* Gets all the tests to pass :)

### Motivation

The reason for doing this part of the migration now is that we are
currently running into occasional issues having to do with the fact that
both the data platform deploy and the Justice Counts team deploy run
migrations against the JC database. This means, depending on the order
in which these deploys are run, we sometimes get into states where the
data platform deploy could fail because it is missing the code for a
migration that was already run as part of the JC deploy. @ohaibbq ran
into this a week or two ago, and it had happened a few times previously
as well.

After this change, JC deploys will run migrations against the new JC
database, and data platform deploys will run migrations against the old
JC database, and so we should no longer have this issue.

### Changes made outside this PR
* Created new SQL instances, one in justice-counts-staging and one in
justice-counts-prod, with the same settings as the instances in
recidiviz-staging and recidiviz-123
* Restored these databases from backups of the corresponding databases
in the recidiviz projects to "initialize" them (followed instructions
[here](https://cloud.google.com/sql/docs/postgres/backup-recovery/restoring#projectid)
to restore to an instance in a different project)
* Added justice_counts_v2_db_*** secrets to all repos. (The reason we
need these in all repos is because the Justice Counts Publisher app
still lives in the classic Recidiviz projects, so we need to be able to
connect to the new DB from those projects too)
* Added a Cloud SQL connection to both JC Cloud Run apps

Yes, I know this should all be done using Terraform. If I had more time
I would TOTALLY learn Terraform. One day, I promise!

### How I tested
* Deployed this PR to a playtesting revision on staging and made sure I
could load both Publisher and Dashboard apps and verified that they are
talking to the new database properly
* Ran `python -m recidiviz.tools.migrations.run_migrations_to_head` to
ensure that my changes to that script worked properly

### Plan for launch
Once this PR is approved, I will:
* Temporarily take down our staging apps (use this [down for
maintenance](https://github.com/Recidiviz/justice-counts/pull/776) page
so no new changes can get in while copying the DB)
* Take a backup of the old staging DB
* Restore the new staging DB from this backup
* Deploy this PR to a playtesting revision on staging and test
* Deploy the PR to 100% traffic on staging
* Repeat the above steps on production

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1106adda1a63c8950e7f9109ea2e91cd62a342ea