Structured logging fix (Recidiviz/recidiviz-data#23331)
## Description of the change

I believe [this
apache-beam](https://github.com/Recidiviz/recidiviz-data/commit/c27dd5f6e90b03a5b3bd765833f098881078a7e5)
upgrade and the corresponding changes to the google-cloud-logging
dependency are causing [this
error](https://recidiviz-inc.sentry.io/issues/4405517999/?project=4504532096516096&query=is%3Aunresolved&referrer=issue-stream&stream_index=0)
for Justice Counts.

This seems to fix, but I'm not sure if it's the "right" solution? It
does unblock us for now though, and if it's possible to get this in at
least temporarily, that would be appreciated!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 94549911b794662a98220c032e38fdd9ced09a4c