Make script for hydrating datetime SQL parsers (Recidiviz/recidiviz-data#23122)
## Description of the change

> As part of a push to deprecate `DATETIME_COL_NORMALIZATION_TEMPLATE`,
we would like all datetime fields to have a SQL parser specified in the
configuration YAML file. This PR introduces a script
`hydrate_datetime_sql_parsers.py` that can be run to automatically fill
in these `datetime_sql_parsers` fields. The script will only label a
column with a parser if all of the non-null values in that column parse
correctly.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | **non-breaking change that adds functionality** |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#12174

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Ryan Guan <rzgn@users.noreply.github.com>
GitOrigin-RevId: 58127e32f76fd3521ef9073cffef9123ee0ad1aa