[Justice Counts] Fix JC admin panel page to connect to the new JC database (Recidiviz/recidiviz-data#23332)
## Description of the change

In https://github.com/Recidiviz/recidiviz-data/pull/23214, I added a
`secret_prefix_override` parameter to a bunch of our DB code, so that we
could connect to the new Justice Counts database (located in the new JC
GCP projects) from applications running in the Recidiviz GCP projects.

I made sure to set this parameter properly for the JC Cloud Run apps,
but forgot to set it appropriately for our page in the Admin Panel, so
that page is now still connecting to our old DB :)

This fixes and plumbs the parameter through a few more paths.

I tested by running the admin panel locally and logging the key of the
secrets it was fetching, to ensure they were "justice_counts_v2"
prefixed.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d3aa6d80e15b5186398bce82074ac44481fd2fdc