[US_MI] Remove erroneous columns from file (Recidiviz/recidiviz-data#23348)
## Description of the change

The MI COMS developers accidentally included 10 erroneous columns in the
table `COMS_Parole_Violation_Summary_Decisions`. Their transfer of the
data from PROD (expected 8/22) will not include these columns so this PR
removes those columns.

Before the deploy with this PR, we should go ahead the deprecate this
table in BQ in:
- [ ] Staging (primary and secondary)
- [ ] Prod (primary and secondary)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a3964958c54b1e5024e6b13806cf88a6a6ecd256