[US_TN] Ingest Violations and Sanctions (Recidiviz/recidiviz-data#23169)
## Description of the change

This PR contains ingest view and mappings from the Violations and
Sanctions raw data tables in TN to hydrate state_supervision_violation
and state_supervision_violation_response (and related entities).

Important to note: This view and mappings is currently gated to staging
only due to a raw data intricacy we are currently debugging. Because of
how TN structures their contact notes which inform many of the
supervision events, we are currently not able to capture all violations
that may lead to reincarceration.

See this thread for a longer write up and discussion about how we may
handle these situations:
https://recidiviz.slack.com/archives/C01J3JV7962/p1692234318184699

For now, we are going to merge in this view so that it can populate
metrics and downstream views which will help in the validation process.
As we learn more from TN folks and the further validations, we will
either refactor this view OR add another view to simulate violations
that are missing in order to inform metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b71f40ec7cc4587630ef4c1dedd88222fb3fc256