[Justice Counts] Actually fix JC admin panel to connect to the new DB (Recidiviz/recidiviz-data#23433)
## Description of the change

In https://github.com/Recidiviz/recidiviz-data/pull/23332 I attempted to
fix the JC Admin Panel page so it would connect to our new databases.
However, I accidentally made the change to the `if in_development()`
block rather than the `if in_gcp()` block ... and so it didn't actually
work. (It did appear to work when I tested, because, well, I was in
development.)

Not sure if there is a good way to _actually_ test this besides
deploying to staging ... ! Once I confirm on staging after today's
deploy, I will cherry-pick into prod.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 490e51b93eb00b4671ecd5af393cf0a6677b5b99