Fix descriptionless columns in raw data Looker dashboards (Recidiviz/recidiviz-data#23451)
## Description of the change

> When a raw file config column has no description, it is not available
in raw data up to date views tables. This wasn't accounted for in the
raw data Looker dashboard generator.
> This PR makes sure that columns without descriptions are treated as
NULL in raw data up to date views tables in Looker dashboards to avoid
query errors.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| **Type: Bug** | **non-breaking change that fixes an issue** |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Ryan Guan <rzgn@users.noreply.github.com>
GitOrigin-RevId: 8b343d9333c2d291882c12e5bd364b1444adba43